Fix missing kdewin depedenceny issue using gpgmepp on Windows.
The KDEWin library is needed on Windows because gpgmepp uses POSIX
types like ssize_t which are not available by default and are provided
by this library of the KDE on Windows project. Ideally gpgmepp should
not use platform dependent types and should solve this the way gpgme
does.

REVIEW:122827