Don't link against multiple gpgmepp flavors.
Using two libraries with exactly the same symbol names but different
implementations is bound to go wrong eventually, and in fact it does,
breaking Kleo unit tests.

The fix on this side is easy, the same however needs to be applied on
the Kleo side, which isn't that straightforward unfortunately.

Options I can think of:
- remove all but the pthread flavor
- put gpgmepp in the public link interface of qgpgme
- duplicate the detection logic on Kleo's side

CCMAIL: Andre Heinecke <aheinecke@intevation.de>