Handle PINENTRY_LAUNCHED status in editinteractor
With GnuPG 2.1 we will get a PINENTRY_LAUNCHED status Kleopatra
would error out on this status as it was unkown.
As the status was new in gpgme 1.5.0 we guard it with an ifdef
to avoid raising the requirement. In older versions this
problem might have been caused by the option allow-pinentry-notify.

This fixes various problems in Kleopatra (adding user id's,
verifying certificates) that would fail if the passphrase/pin was
not cached.

CCBUG: 353252