Do not treat KEYEXPIRED as error
GnuPG doc/DETAILS already mentions that this status code
is of limited usefulness as it is emited as soon as one
subkey is expired. So this can't be used as an error indication.

Backport of GpgME Commit: 82d484c8

BUG: 358392

Merge: None. This needs rev. 3872dcaa to actually make the commands
work.