Handle gpgme_data_identify properly
This reduces the dependency back again to 1.1.7 as it
adds a check for data_identify. Additionally it checks for
the new binary and more fingrained version of gpgme_data_identify
that can actually be used and provides a feature check for
this.

This is a backport of GpgME commit: 15fc5c34