Handle empty recipients consistently
d6f5f692 was incomplete as the code to handle recpients
was duplicated four times. This is now unified and constently
uses a nullptr instead of an empty array.

As this does not really fix any working feature in GpgME++
it is only commited in master because there is a slight regression
risk and it adds a new (private) symbol in context.