Use gpgme_data_set_flag to set filesize
The dataprovider class and memory functions did not have a file
size knowable to gnupg. With GnuPG 2.1.15 we will get a size-hint
that can be used to tell gnupg the size to expect of streams.

We use that to tell gnupg for Seekable DataProviders and memory
functions which size to expect.

This has the effect of fixing the Progress handling in QGpgME.